update: mariadb sql to MS server conversion
updated general sql syntax to MS server with conditional db drop & Transactions marked with go.

+++ Day15/employees.sql
@@ -1,8 +1,15 @@
-DROP DATABASE EMP;
+USE master;
+GO
+
+IF EXISTS (SELECT name FROM master.dbo.sysdatabases WHERE name = 'EMP')
+    DROP DATABASE EMP;
+GO
 
 CREATE DATABASE EMP;
+GO
 
 USE EMP;
+GO
 
 CREATE TABLE ITEM (
     itemname VARCHAR(50) PRIMARY KEY,
